+++ src/audio.rs
@@ -27,7 +27,7 @@ pub trait Device: fmt::Display + std::marker::Send + std::marker::Sync {
     fn play(
         &self,
         song: Arc<Song>,
-        mappings: &HashMap<String, u16>,
+        mappings: &HashMap<String, Vec<u16>>,
         cancel_handle: CancelHandle,
         play_barrier: Arc<Barrier>,
     ) -> Result<(), Box<dyn Error>>;